Move feedback feeling card to training page
