+++ lib/pages/trainee_feedback.dart
@@ -3,6 +3,7 @@ import 'package:intl/intl.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 import '../l10n/app_localizations.dart';
+import 'training.dart' show FeedbackFeelingCard, FeelingOption;
 
 class TraineeFeedbackPage extends StatefulWidget {
   const TraineeFeedbackPage({super.key});
@@ -63,7 +64,7 @@ class _TraineeFeedbackPageState extends State<TraineeFeedbackPage> {
               minLines: 4,
             ),
             const SizedBox(height: 12),
-            _FeedbackFeelingCard(
+            FeedbackFeelingCard(
               label: l10n.traineeFeedbackFeelingLabel,
               helperText: l10n.traineeFeedbackFeelingHint,
               selectedFeeling: _selectedFeeling,
@@ -375,7 +376,7 @@ class _FeedbackCard extends StatelessWidget {
     final theme = Theme.of(context);
     final dateText = DateFormat.yMMMd(l10n.localeName).format(feedback.createdAt);
     final isRead = feedback.readAt != null;
-    final feelingOption = _FeelingOption.fromValue(feedback.feeling);
+    final feelingOption = FeelingOption.fromValue(feedback.feeling);
 
     return Card(
       margin: const EdgeInsets.only(bottom: 12),
@@ -426,104 +427,5 @@ class _FeedbackCard extends StatelessWidget {
         ),
       ),
     );
-  }
-}
-
-class _FeedbackFeelingCard extends StatelessWidget {
-  final String label;
-  final String helperText;
-  final int? selectedFeeling;
-  final ValueChanged<int> onSelected;
-
-  const _FeedbackFeelingCard({
-    required this.label,
-    required this.helperText,
-    required this.selectedFeeling,
-    required this.onSelected,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    final theme = Theme.of(context);
-
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              label,
-              style: theme.textTheme.titleMedium?.copyWith(
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            const SizedBox(height: 8),
-            Text(
-              helperText,
-              style: theme.textTheme.bodyMedium?.copyWith(
-                color: theme.colorScheme.onSurfaceVariant,
-              ),
-            ),
-            const SizedBox(height: 12),
-            Wrap(
-              spacing: 8,
-              runSpacing: 8,
-              children: _FeelingOption.values
-                  .map(
-                    (option) => ChoiceChip(
-                      selected: selectedFeeling == option.value,
-                      onSelected: (_) => onSelected(option.value),
-                      label: Text(
-                        '${option.emoji} ${option.localizedLabel(l10n)}',
-                      ),
-                    ),
-                  )
-                  .toList(growable: false),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _FeelingOption {
-  final int value;
-  final String emoji;
-
-  const _FeelingOption({required this.value, required this.emoji});
-
-  static const veryBad = _FeelingOption(value: 1, emoji: '😣');
-  static const bad = _FeelingOption(value: 2, emoji: '😕');
-  static const ok = _FeelingOption(value: 3, emoji: '😐');
-  static const good = _FeelingOption(value: 4, emoji: '🙂');
-  static const veryGood = _FeelingOption(value: 5, emoji: '🤩');
-
-  static const values = [veryBad, bad, ok, good, veryGood];
-
-  static _FeelingOption? fromValue(int? value) {
-    for (final option in values) {
-      if (option.value == value) return option;
-    }
-    return null;
-  }
-
-  String localizedLabel(AppLocalizations l10n) {
-    switch (value) {
-      case 1:
-        return l10n.traineeFeedbackFeelingVeryBad;
-      case 2:
-        return l10n.traineeFeedbackFeelingBad;
-      case 3:
-        return l10n.traineeFeedbackFeelingOk;
-      case 4:
-        return l10n.traineeFeedbackFeelingGood;
-      case 5:
-        return l10n.traineeFeedbackFeelingVeryGood;
-      default:
-        return '';
-    }
   }
 }
